Group saving

+++ app/src/main/java/org/grassroot/android/utils/RealmUtils.java
@@ -98,10 +98,12 @@ public class RealmUtils {
       String pValue) {
     Realm realm = Realm.getDefaultInstance();
     realm.beginTransaction();
-    realm.where(clazz).equalTo(pName, pValue).findFirst().deleteFromRealm();
+    RealmObject object = realm.where(clazz).equalTo(pName, pValue).findFirst();
+    if(object!=null) object.deleteFromRealm();
     realm.commitTransaction();
     realm.close();
   }
+
 
   public static void removeObjectsFromDatabase(Class<? extends RealmObject> clazz,
       Map<String, Object> map) {
@@ -115,7 +117,7 @@ public class RealmUtils {
       }
     }
     realm.beginTransaction();
-    query.findAll().deleteAllFromRealm();
+    if(query.findAll().size()>0) query.findAll().deleteAllFromRealm();
     realm.commitTransaction();
     realm.close();
   }
